Tests/LibWeb/Text/expected/hit_testing/inline-content-inside-anonymous-container.txt

<#text>
index: 4
<BODY>
---
<#text>
index: 20
<A id="target">
---
<DIV id="a">
index: 0
<BODY>
---
